Move EvenrReader out of sdxl_turbo module
