+++ src/commands/ai/hugging_face.ts
@@ -1,4 +1,4 @@
-export interface EventEstimation {
+interface EventEstimation {
     msg: 'estimation';
     rank: number;
     queue_size: number;
@@ -7,22 +7,22 @@ export interface EventEstimation {
     rank_eta: number;
     queue_eta: number;
 }
-export interface EventSendData {
+interface EventSendData {
     msg: 'send_data';
     event_id: string;
 }
-export interface EventProcessStart {
+interface EventProcessStart {
     msg: 'process_starts';
 }
-export interface EventProcessCompleted {
+interface EventProcessCompleted {
     msg: 'process_completed';
     output?: Output;
     success: boolean;
 }
 
-export type Event = EventEstimation | EventSendData | EventProcessStart | EventProcessCompleted;
+type Event = EventEstimation | EventSendData | EventProcessStart | EventProcessCompleted;
 
-export interface Input {
+interface Input {
     session_hash: string;
 
     prompt: string;
@@ -31,14 +31,14 @@ export interface Input {
     seed: number;
 }
 
-export interface Output {
+interface Output {
     data: OutputData[];
     is_generating: boolean;
     duration: number;
     average_duration: number;
 }
 
-export interface OutputData {
+interface OutputData {
     path: string;
     url: string | null;
     size: number | null;
@@ -46,7 +46,7 @@ export interface OutputData {
     mime_type: string | null;
 }
 
-export async function send_data<T>(event_id: string, session_hash: string, data: T): Promise<boolean> {
+async function send_data<T>(event_id: string, session_hash: string, data: T): Promise<boolean> {
     const res = await fetch('https://diffusers-unofficial-sdxl-turbo-i2i-t2i.hf.space/queue/data', {
         method: 'POST',
         body: JSON.stringify({
@@ -62,4 +62,125 @@ export async function send_data<T>(event_id: string, session_hash: string, data:
         },
     });
     return res.status == 200;
+}
+
+interface Attachment {
+    name: string;
+    attachment: Buffer;
+}
+
+async function getRoot(): Promise<Option<string>> {
+    const res = await fetch('https://diffusers-unofficial-sdxl-turbo-i2i-t2i.hf.space/?__theme=light', {
+        method: 'GET',
+    });
+    if (res.status != 200) {
+        return null;
+    }
+    const html = await res.text();
+
+    const reg = /(https:\/\/diffusers-unofficial-sdxl-turbo-i2i-t2i.hf.space\/--replicas\/\w+)/s.exec(html);
+    if (!reg?.[1]) {
+        return null;
+    }
+    return reg[1];
+}
+
+let ROOT: Option<string> = await getRoot();
+
+async function getFileFromRoot(path: string, force: boolean = true): Promise<Option<ArrayBuffer>> {
+    if (!ROOT) {
+        await getRoot();
+        if (!ROOT) {
+            return null;
+        }
+    }
+    const res = await fetch(`${ROOT}/file=${path}`);
+    if (res.status == 404 && !force) {
+        ROOT = null;
+        return getFileFromRoot(path, false);
+    } else if (res.status != 200) {
+        return null;
+    }
+    return res.arrayBuffer();
+}
+
+function intoEvent(value_string: string): Event | null {
+    const reg = /data: (.*)/.exec(value_string);
+    if (!reg?.[1]) {
+        return null;
+    }
+    const data = reg[1];
+    const parsed = JSON.parse(data) as Event;
+    return parsed;
+}
+
+export class EventReader {
+    private img: Option<Attachment> = null;
+    public constructor(
+        private reader: ReadableStreamDefaultReader<Uint8Array>,
+        private data: Input,
+    ) {}
+    public image(): Option<typeof this.img> {
+        return this.img;
+    }
+
+    public async process(): Promise<void> {
+        for (;;) {
+            const evt = await this.reader.read();
+            if (evt.done) {
+                return;
+            }
+            const value_string = new TextDecoder('utf-8').decode(evt.value);
+            for (const line of value_string.split('\n')) {
+                if (line === '') {
+                    continue;
+                }
+                const evt = intoEvent(line);
+                if (!evt) {
+                    continue;
+                }
+                if (!(await this.processEvent(evt))) {
+                    return;
+                }
+            }
+        }
+    }
+    private async processEvent(evt: Event): Promise<boolean> {
+        switch (evt.msg) {
+            // case "estimation":
+            //     break;
+            case 'send_data':
+                if (
+                    !(await send_data(evt.event_id, this.data.session_hash, [
+                        null,
+                        this.data.prompt,
+                        this.data.strength,
+                        this.data.steps,
+                        this.data.seed,
+                    ]))
+                ) {
+                    return false;
+                }
+                break;
+            // case "process_starts":
+            //     break;
+            case 'process_completed':
+                if (evt.success) {
+                    const data = evt.output?.data[0];
+                    if (!data) {
+                        return false;
+                    }
+                    const res = await getFileFromRoot(data.path);
+                    if (!res) {
+                        return false;
+                    }
+                    this.img = {
+                        name: data.orig_name,
+                        attachment: Buffer.from(res),
+                    };
+                }
+                break;
+        }
+        return true;
+    }
 }
